:bug: Examplar data moved out of tests
